Pulled out form from resource page.
To be consistent with other forms on the site, this form has now become it's own component.
