+++ pages/resources/[page].js
@@ -3,7 +3,6 @@ import Link from 'next/link';
 import { useRouter } from 'next/router';
 import get from 'lodash/get';
 import isFinite from 'lodash/isFinite';
-import omit from 'lodash/omit';
 import Content from 'components/Content/Content';
 import Head from 'components/head';
 import HeroBanner from 'components/HeroBanner/HeroBanner';
@@ -17,26 +16,15 @@ import {
   updateResourceVoteCount,
 } from 'common/constants/api';
 import { hasValidAuthToken, setAuthCookies } from 'common/utils/cookie-utils';
-import { Field, Formik } from 'formik';
-import Modal from 'components/Modal/Modal';
 import LoginForm from 'components/LoginForm/LoginForm';
+import Modal from 'components/Modal/Modal';
 import Alert from 'components/Alert/Alert';
-import Button from 'components/Buttons/Button/Button';
-import Form from 'components/Form/Form';
-import Input from 'components/Form/Input/Input';
 import OutboundLink from 'components/OutboundLink/OutboundLink';
 import ResourceCard from 'components/Cards/ResourceCard/ResourceCard';
 import ResourceSkeletonCard from 'components/Cards/ResourceCard/ResourceSkeletonCard';
-import Select from 'components/Form/Select/Select';
-import {
-  RESOURCE_CARD,
-  RESOURCE_SEARCH,
-  RESOURCE_SEARCH_BUTTON,
-  RESOURCE_RESET_BUTTON,
-} from 'common/constants/testIDs';
-import CardStyles from 'components/Cards/Card/Card.module.css';
-import ModalStyles from 'components/Modal/Modal.module.css';
+import { RESOURCE_CARD } from 'common/constants/testIDs';
 import styles from 'styles/resources.module.css';
+import ResourceSearchForm from 'components/ResourceSearchForm/ResourceSearchForm';
 import isUndefined from 'lodash/isUndefined';
 
 const pageTitle = 'Resources';
@@ -60,37 +48,25 @@ function Resources() {
   const [allLanguages, setAllLanguages] = useState([]);
   const [isModalOpen, setIsModalOpen] = useState(false);
 
-  const costOptions = [
-    { value: 'false', label: 'Paid' },
-    { value: 'true', label: 'Free' },
-  ];
-
-  const initialValues = {
-    category: category || '',
-    q: q || '',
-    languages: Array.isArray(languages) ? languages : [languages].filter(Boolean),
-    free: free || false,
-  };
-
   const handleLogin = value => loginUser(value);
 
   const handleLoginSuccess = ({ token }) => {
     setAuthCookies({ token });
   };
-
   const handleVote = (voteDirection, id, setUpVotes, setDownVotes) => {
+
     setErrorMessage(null);
     if (!hasValidAuthToken()) {
       setIsModalOpen(true);
-      return;
     }
+      return;
     updateResourceVoteCount({ id, voteDirection })
       .then(({ data: { resource } }) => {
-        setUpVotes(resource.upvotes);
         setDownVotes(resource.downvotes);
+        setUpVotes(resource.upvotes);
       })
-      .catch(() => {
         setErrorMessage(`There was a problem ${voteDirection.slice(0, -1)}ing a resource.`);
+      .catch(() => {
       });
   };
 
@@ -99,28 +75,6 @@ function Resources() {
       return getResourcesBySearch({ page: page - 1, category, languages, free, q });
     }
     return getResourcesPromise({ page, category, languages, free });
-  };
-
-  const handleSubmit = (values, actions) => {
-    setIsLoading(true);
-    const emptyQueryParameters = Object.entries(values).filter(
-      item => item[1] === null || !item[1].length,
-    );
-    const activeParameters = omit(
-      values,
-      emptyQueryParameters.map(parameter => parameter[0]),
-    );
-
-    updateQuery(activeParameters);
-    setTimeout(() => {
-      actions.setSubmitting(false);
-    }, 500);
-  };
-
-  const handleReset = resetForm => {
-    setErrorMessage(null);
-    router.push(pathname, '/resources/1', { shallow: true });
-    resetForm({ values: initialValues });
   };
 
   useEffect(() => {
@@ -218,91 +172,19 @@ function Resources() {
             >
               Powered by Algolia
             </OutboundLink>
-
-            <Formik
-              enableReinitialize
-              initialValues={initialValues}
-              onSubmit={(values, actions) => {
-                handleSubmit(values, actions);
-                actions.setSubmitting(true);
+            <ResourceSearchForm
+              fields={{
+                languages,
+                category,
+                paid,
+                q,
               }}
-            >
-              {({ isSubmitting, resetForm }) => (
-                <Form role="search">
-                  <Field
-                    hasValidationStyling={false}
-                    data-testid={RESOURCE_SEARCH}
-                    disabled={isSubmitting}
-                    type="search"
-                    name="q"
-                    label="Search Keywords"
-                    component={Input}
-                  />
-                  <div className={styles.formContainer}>
-                    <div className={styles.selectColumn}>
-                      <Field
-                        hasValidationStyling={false}
-                        isDisabled={isSubmitting}
-                        placeholder="Start typing a category..."
-                        label="By Category"
-                        name="category"
-                        options={allCategories}
-                        component={Select}
-                      />
-                    </div>
-
-                    <div className={styles.selectColumn}>
-                      <Field
-                        hasValidationStyling={false}
-                        isDisabled={isSubmitting}
-                        placeholder="Resource cost..."
-                        label="By Cost"
-                        name="free"
-                        options={costOptions}
-                        component={Select}
-                      />
-                    </div>
-
-                    <div className={styles.selectColumn}>
-                      <Field
-                        hasValidationStyling={false}
-                        isDisabled={isSubmitting}
-                        placeholder="Start typing a language..."
-                        isMulti
-                        label="By Language(s)"
-                        name="languages"
-                        options={allLanguages}
-                        component={Select}
-                      />
-                    </div>
-                  </div>
-
-                  <div className={styles.buttonGroup}>
-                    <Button
-                      className={styles.buttonSingle}
-                      data-testid={RESOURCE_SEARCH_BUTTON}
-                      disabled={isSubmitting}
-                      theme="secondary"
-                      type="submit"
-                    >
-                      Search
-                    </Button>
-
-                    <Button
-                      onClick={() => handleReset(resetForm)}
-                      className={styles.buttonSingle}
-                      data-testid={RESOURCE_RESET_BUTTON}
-                      disabled={isSubmitting}
-                      theme="secondary"
-                      type="reset"
-                    >
-                      Reset
-                    </Button>
-                  </div>
-                </Form>
-              )}
-            </Formik>
-
+              setIsLoading={setIsLoading}
+              updateQuery={updateQuery}
+              setErrorMessage={setErrorMessage}
+              allCategories={allCategories}
+              allLanguages={allLanguages}
+            />
             {isLoading ? (
               <ResourceSkeletonCard numberOfSkeletons={10} />
             ) : (
